+++ src/components/project-selector/project-selector.vue
@@ -1,0 +1,2 @@
+<template src="./project-selector.html" />
+<script src="./project-selector.ts" lang="ts" />