+++ src/router/index.ts
@@ -6,32 +6,37 @@ import { Auth0 } from '../auth'
 export const ROUTES_NAME = Object.freeze({
   root: 'root',
   overview: 'overview',
-  species_richness: 'species_richness'
+  species_richness: 'species_richness',
+  error: 'error'
 })
 
 const routes: RouteRecordRaw[] = [
   {
-    path: '/',
+    path: '/project/:projectId',
     component: Pages.RootPage,
     beforeEnter: Auth0.routeGuard,
     children: [
       {
-        path: '/',
+        path: '/project/:projectId/',
         redirect: {
           name: ROUTES_NAME.overview
         }
       },
       {
-        path: '/overview',
+        path: '/project/:projectId/overview',
         name: ROUTES_NAME.overview,
         component: Pages.OverviewPage
       },
       {
-        path: '/species_richness',
+        path: '/project/:projectId/species_richness',
         name: ROUTES_NAME.species_richness,
         component: Pages.SpeciesRichnessPage
       }
     ]
+  },
+  {
+    path: '/:pathMatch(.*)*',
+    component: Pages.ErrorPage
   }
 ]