+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * **setup:** CE-1261 Setup biodiversity analytics project ([CE-1261](https://jira.rfcx.org/browse/CE-1261))
 * **setup:** CE-1297 Setup biodiversity analytics deployment ([CE-1297](https://jira.rfcx.org/browse/CE-1297))
+* **setup:** CE-1335 Setup navigation bar with project selector ([CE-1335](https://jira.rfcx.org/browse/CE-1335))
 * **setup:** CE-1262 Set up authentication page ([CE-1262](https://jira.rfcx.org/browse/CE-1262))
 * **setup:** CE-1348 Set up biodiversity analytics store data tool([CE-1348](https://jira.rfcx.org/browse/CE-1348))
 * **setup:** Set up api service for calling an api